Actors

An Actor is a computational entity that, in response to a message it receives, can concurrently:

send a finite number of messages to other actors;

create a finite number of new actors;

designate the behavior to be used for the next message it receives. [1]

YAActL provides a library for messages, links (message channels), behaviors and actors which enables us to implement actor systems in Julia.

Start

Actor starts an actor on a behavior and returns a link to it.

YAActL.Actor — Function

Actor([lp::LinkParams], bhv::Function, args...; kwargs...)
Actor(pid::Int, bhv::Function, args...; kwargs...)

Create a new actor. Start a task executing repeatedly the behavior bhv. The actor listens to messages msg sent over the returned link and executes bhv(args..., msg, kwargs) for each message. The actor stops if sent Stop().

Arguments

[lp::LinkParams]: optional parameters for creating the actor,

pid::Int: create the actor on process pid,

bhv::Function: function implementing the actor's behavior,

args...: arguments to bhv, (without msg)

kwargs...: keyword arguments to bhv.

Returns

a Link to a locally created actor or

an RLink to a remote actor.

Example

In the following code snippet we start an actor with a stack_node behavior function and one of its arguments (an empty StackNode).

mystack = Actor(stack_node, StackNode(nothing, Link()))

We saved the returned link in mystack and can now send! messages to it.

Operation

If an actor gets a message, it executes its behavior function with it.

Actors can switch their behavior function with become or we can cause them to switch it with become!. Thereby an actor can represent a state machine.

We can stop an Actor.

YAActL.stopActor — Function

stopActor(): an actor terminates.

YAActL.stopActor! — Function

stopActor!(lk::LINK): terminate an actor with link lk.

1See: The Actor Model on Wikipedia.