Actor API

Installation

YAActL — Module

YAActL

Yet Another Actor Library (in Julia VERSION ≥ v"1.3").

The current stable, registered version is installed with

pkg> add YAActL

The development version is installed with:

pkg> add("https://github.com/pbayer/YAActL.jl")

YAActL.version — Constant

Gives the package version.

julia> using YAActL

julia> YAActL.version
v"0.1.2"

The following functions provide a user interface to actors:

Start

YAActL.Actor — Function

Actor([lp::LinkParams], bhv::Function, args...; kwargs...)
Actor(pid::Int, bhv::Function, args...; kwargs...)

Create a new actor. Start a task executing repeatedly the behavior bhv. The actor listens to messages msg sent over the returned link and executes bhv(args..., msg, kwargs) for each message. The actor stops if sent Stop().

Arguments

[lp::LinkParams]: optional parameters for creating the actor,

pid::Int: create the actor on process pid,

bhv::Function: function implementing the actor's behavior,

args...: arguments to bhv, (without msg)

kwargs...: keyword arguments to bhv.

Returns

a Link to a locally created actor or

an RLink to a remote actor.

Links

An actor is only represented by its link which it returns upon creation:

YAActL.Link — Type

A Channel{Message} type for communicating with local actors.

Use buffered channels!

In actor systems we always use buffered links to avoid blocking. Responding on an unbuffered or full link causes an actor to block. Link() creates an unbuffered Channel, use Link(32) or newLink() instead!

Example

julia> response = Link(32)
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.RLink — Type

A RemoteChannel{Link} type for communicating with remote actors.

YAActL.LINK — Constant

A Union{Link, RLink} type for communicating with actors.

Actors correspond with other actors over links. There is a default link for users to communicate with actors.

YAActL.USR — Constant

User remote channel for interacting with actors.

For setting up links explicitly we have the following functions.

YAActL.newLink — Function

newLink(sz::Integer=32)

Create a link of buffer size sz to get responses from actors. Buffer sizes sz < 10 are not allowed.

Example

julia> response = newLink()
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.LinkParams — Type

LinkParams(pid=myid(), size=32; taskref=nothing, spawn=false)

Parameters for setting up an Actor.

pid::Int: process identification,

size::Int: channel buffer size, must be size ≥ 10,

taskref::Union{Nothing, Ref{Task}}: If you need a reference to the created task, pass a Ref{Task} object via the keyword argument taskref.

spawn::Bool: If spawn = true, the Task created may be scheduled on another thread in parallel, equivalent to creating a task via Threads.@spawn.

YAActL.parallel — Function

parallel(size=32; taskref=nothing)

Return LinkParams with spawn=true.

Example

julia> using YAActL, .Threads

julia> myactor = Actor(parallel(), threadid);

julia> call!(myactor)
2

We send messages to actors and they can send them to others over links.

Messages

Messages in YAActL have a common abstract type. Only a few predefined messages are exported:

YAActL.Message — Type

Abstract type for messages to actors.

YAActL.Response — Type

Response(y, from::LINK=self())

A Message representing a response to requests.

YAActL.Request — Type

Request(x, from::LINK)

A generic Message for user requests.

YAActL.Timeout — Type

Timeout()

A return value to signal that a timeout has occurred.

Actors operate with internal messages. Further messages can be implemented by a user. If an actor receives a message other than an internal one, it passes the message as last argument to the behavior function.

Send and Receive

Messages are sent and received using the following basic functions:

YAActL.send! — Function

send!(lk::LINK, m::Message)

Send a message m to an actor over a LINK lk.

send!(lks::Tuple{LINK,Vararg{LINK}}, m::M) where M<:Message
send!(lks::Vector{LINK}, m::M) where M<:Message

Send a message m to a Vector or Tuple of LINKs.

YAActL.receive! — Function

receive!(lk; timeout=5.0)
receive!(lk, from; timeout=5.0)
receive!(lk, Msg; timeout=5.0)
receive!(lk, Msg, from; timeout=5.0)

Receive a message over a link lk.

If Msg or from are provided, receive! returns only a matching message.

Parameters

lk::LINK: local or remote link over which the message is sent,

Msg::Type{<:Message}: Message type,

from::LINK: local or remote link of sender. If from is provided, only messages with a from field can be matched.

timeout::Real: maximum waiting time in seconds. If timeout==0, lk is scanned for existing messages. Set timeout=Inf if you don't want receive! to timeout.

Returns

received message or Timeout().

Actor Control

The following functions control actor behavior and state by sending implicit messages. To those actors don't send response messages.

YAActL.become! — Function

become!(lk::LINK, bhv::Function, args...; kwargs...)

Cause an actor to switch its behavior.

Arguments

lk::Link: Link to an actor,

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.cast! — Function

cast!(lk::LINK, args...)

Execute the lk actor's behavior function with args... without sending a response. The returned value of the behavior function can be obtained with query!

YAActL.exit! — Function

exit!(lk::LINK, code=0)

Tell an actor lk to exit. If it has a term function it calls it with code as last argument.

This is an asynchronous message without a response.

YAActL.init! — Function

init!(lk::LINK, f::Function, args...; kwargs...)

Tell an actor lk to execute the init function f with the given arguments at startup and to save the returned value as state variable.

YAActL.update! — Function

update!(lk::LK, args...)

Update the lk actor's internal state with args.... This is an asynchronous message without a response.

It can be called with Args to update the stored arguments to the behavior function. If Args has keyword arguments, they are merged with existing keyword arguments to the behavior function.

YAActL.set! — Function

set!(lk::LINK, dsp::Dispatch)

Set the lk actor's Dispatch to dsp.

YAActL.term! — Function

term!(lk::LINK, f::Function, args...; kwargs...)

Tell an actor lk to execute a function f when it terminates. f must accept a code=0 as last argument. This is added by the actor to args... when it exit!s.

If a behavior function wants to control its own actor, it can use the following functions:

YAActL.become — Function

become(bhv::Function, args...; kwargs...)

Cause your actor to take on a new behavior. This can only be called from inside an actor/behavior.

Arguments

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.self — Function

self()

Get the LINK of your actor.

YAActL.stop — Function

stop(code=0)

Cause your actor to exit with code.

Bidirectional Messaging

The following functions send messages to actors causing them to send a Response. There are two ways to do it:

Send a message with an explicit from-link to an actor and it will respond to it. Then you can asynchronously receive! the response.

Send a message with an implicit link to an actor, block and wait synchronously for the response.

There is a basic function for synchronous communication:

YAActL.request! — Function

request!(lk::LINK, msg::Message; full=false, timeout::Real=5.0)
request!(lk::LINK, Msg::Type{<:Message}, args...; kwargs...)

Send a message to an actor, block, receive and return the result.

Arguments

lk::LINK: actor link,

msg::Message: a message,

Msg::Type{<:Message}: a message type,

args...: optional arguments to Msg,

full: if true return the full Response message.

timeout::Real=5.0: timeout in seconds after which a Timeout is returned,

kwargs...: full or timeout.

The following functions operate

asynchronously if you provide them with a from-link or

synchronously (they block) if you don't.

YAActL.call! — Function

call!(lk::LINK, from::LINK, args...)
call!(lk::LINK, args...; timeout::Real=5.0)

Call the lk actors behavior function withargs...and send its returned value as [Response](@ref) to thefromchannel. Iffromis omitted,call!` blocks and returns the response.

Examples

Base.get! — Function

get!(lk::LINK, from::LINK)
get!(lk::LINK; timeout::Real=5.0)

Get a Response message from the lk actor with its internal state to the from channel. If from is omitted, get! blocks and returns the response.

YAActL.exec! — Function

exec!(lk::LINK, from::LINK, f::Function, args...; kwargs...)
exec!(lk::LINK, from::LINK, fu::Func)
exec!(lk::LINK, fu::Func; timeout::Real=5.0)

Tell an actor to execute an arbitrary function and to send the returned value as a Response.

Arguments

lk::LINK: link to the actor,

from::LINK: the link a Response should be sent to. If from is ommitted, exec! does a synchronous call and returns the result. In that case there is a timeout.

f::Function, args...; kwargs... or

fu::Func: function arguments,

timeout::Real=5.0: timeout in seconds.

to the from channel. If from is omitted, exec! blocks and returns the response value.

YAActL.query! — Function

query!(lk::LINK, [from::LINK])
query!(lk::LINK; timeout::Real=5.0)

Query the result of the last call to the behavior function from the lk actor. The Response is sent to the from channel. If from is omitted query! blocks and returns the response.