Diagnosis

In order to develop actor programs, it is useful to have access to the actor tasks and eventually to their stack traces. You can register! an actor channel to a Vector{Link} in order to get access to the tasks.

YAActL.register! — Function

register!(lks::Vector{Link}, lk::Link)

Register a link lk to a vector of links lks.

Base.istaskfailed — Method

istaskfailed(lk::Link)

Returns true if a task associated with lk has failed.

Base.istaskfailed — Method

istaskfailed(lks::Vector{Link})

Returns true if any task associated with a vector lks of links has failed.

YAActL.info — Function

info(lk::Link)

Return the state (eventually the stacktrace) of a task associated with lk.

For diagnostic purposes it is possible to get access to the actor's ACT variable:

YAActL.diag! — Function

diag!(lk::LINK)

Return the internal _ACT variable of the lk actor. This is only for diagnosis and testing.