Links

We send messages to actors and they can send them to others over links. In fact, an actor is only represented by its link which it returns upon creation:

YAActL.Link — Type

A Channel{Message} type for communicating with local actors.

Warn

In actor systems we always use buffered links to avoid blocking. Responding on an unbuffered or full link causes an actor to block. Link() creates an unbuffered Channel, use Link(32) or newLink() instead!

Example

julia> response = Link(32)
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.RLink — Type

A RemoteChannel{Link} type for communicating with remote actors.

YAActL.LINK — Constant

A Union{Link, RLink} type for communicating with actors.

YAActL.send! — Function

send!(lk::LINK, m::Message)

Send a message m to an actor over a LINK lk.

send!(lks::Tuple{LINK,Vararg{LINK}}, m::M) where M<:Message
send!(lks::Vector{LINK}, m::M) where M<:Message

Send a message m to a Vector or Tuple of LINKs.

Actors correspond with other actors over links. If we want a response from an actor, we must send it our own link together with a request message.

YAActL.self — Method

self()

Get a local Link to yourself from inside an actor.

YAActL.USR — Constant

User channel for interacting with actors.

For setting up links explicitly we have the following functions.

YAActL.newLink — Function

newLink(sz::Integer=32)

Create a link of buffer size sz to get responses from actors. Buffer sizes sz < 10 are not allowed.

Example

julia> response = newLink()
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.LinkParams — Type

LinkParams(size=32; taskref=nothing, spawn=false)

Set the parameters for setting up an Actor.

Parameters

pid::Int: process identification,

size::Int: channel buffer size, must be size ≥ 10,

taskref::Union{Nothing, Ref{Task}}: If you need a reference to the created task, pass a Ref{Task} object via the keyword argument taskref.

spawn::Bool: If spawn = true, the Task created may be scheduled on another thread in parallel, equivalent to creating a task via Threads.@spawn.

persistent::Bool: if persistent = false, the

YAActL.parallel — Function

parallel(size=32; taskref=nothing)

Return LinkParams with spawn=true.