Messages

Messages in YAActL have a common abstract type. This enables actors to dispatch their behavior functions on message types.

Some basic message types are for setting up the type hierarchy and for controlling the actors themselves:

YAActL.Message — Type

Abstract type for messages to actors.

YAActL.Stop — Type

Stop()

A message causing an actor to stop.

YAActL.Become — Type

Become(f::Function, args...; kwargs...)

Internal message for behavior change.

The following message types are for executing and dispatching standard behaviors:

YAActL.Request — Type

Request(x, lk::LINK)

A message representing a request of x by customer lk.

YAActL.Response — Type

Response(y)

A message representing a response of y to a request.

Other message types can be implemented by the user, for example:

struct Pop <: Message
    customer::Link
end

struct Push{T} <: Message
    content::T
end

With dispatch on message types we can easily implement state machines.