Usage

An Actor is a computational entity that, in response to a message it receives, can concurrently:

send a finite number of messages to other actors;

create a finite number of new actors;

designate the behavior to be used for the next message it receives. [1]

YAActL provides a library for messages, links (message channels), behaviors and actors which enables us to implement actor systems in Julia.

Messages

Messages in YAActL have a common abstract type. This enables actors to dispatch their behavior functions on message types.

Some basic message types are for setting up the type hierarchy and for controlling the actors themselves:

YAActL.Message — Type

Abstract type for messages to actors.

YAActL.Stop — Type

Stop(): a message causing an actor to stop.

YAActL.Become — Type

Become(f::Function, args...; kwargs...): internal message for behavior change.

The following message types are for executing and dispatching standard behaviors:

YAActL.Request — Type

Request(x, u::Link): a message representing a request of x by customer u.

YAActL.Response — Type

Response(y): a message representing a response of y to a request.

Other message types can be implemented by the user, for example:

struct Pop <: Message
    customer::Link
end

struct Push{T} <: Message
    content::T
end

With dispatch on message types we can easily implement state machines.

Links

We send messages to actors and they can send them to others over links. In fact, an actor is only represented by its link. If we want a response from an actor, we must send it our own link together with a request message.

YAActL.Link — Type

A Channel{Message} type for communicating with actors.

Warn

In actor systems we always use buffered links to avoid blocking. Responding on an unbuffered or full link causes an actor to block. Link() creates an unbuffered Channel, use Link(32) or newLink() instead!

Example

julia> response = Link(32)
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.newLink — Function

newLink(sz::Integer=32)

Create a link of buffer size sz to get responses from actors. Buffer sizes sz < 10 are not allowed.

Example

julia> response = newLink()
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.LinkParams — Type

LinkParams(size=32; taskref=nothing, spawn=false)

Set the parameters for setting up an Actor.

Parameters

size::Int: channel buffer size, must be size ≥ 10,

taskref::Union{Nothing, Ref{Task}}: If you need a reference to the created task, pass a Ref{Task} object via the keyword argument taskref.

spawn::Bool: If spawn = true, the Task created may be scheduled on another thread in parallel, equivalent to creating a task via Threads.@spawn.

persistent::Bool: if persistent = false, the

YAActL.parallel — Function

parallel(size=32; taskref=nothing)

Return LinkParams with spawn=true.

Behaviors

When a message arrives, the actor executes a behavior function on it. Therefore the behavior function must take a Message as its last argument.

In the following example we define two behaviors forward! and stack_node. There are two methods for stack_node, dispatching on Push and Pop.

forward!(lk::L, msg::M) where {L<:Link, M<:Message} = send!(lk, msg)

function stack_node(sn::StackNode, msg::Pop)
    isnothing(sn.content) || become(forward!, sn.link)
    send!(msg.customer, Response(sn.content))
end

function stack_node(sn::StackNode, msg::Push)
    P = Actor(stack_node, sn)
    become(stack_node, StackNode(msg.content, P))
end

Actors change their behavior with become. They also can create other actors. They can send messages to themselves and other actors.

Actors

Actor starts an actor on a behavior and returns a link to it.

mystack = Actor(stack_node, StackNode(nothing, Link()))

We pass the behavior function and its arguments (an empty StackNode but without the last msg argument!) to the actor.

Over the returned link we can send! messages to it.

YAActL.Actor — Function

Actor([lp::LinkParams], bhv::Function, args...; kwargs...)

Create a new actor. Start a task executing repeatedly the behavior bhv. The actor listens to messages msg sent over the returned link and executes bhv(args..., msg, kwargs) for each message. The actor stops if sent Stop().

Arguments

[lp::LinkParams]: optional parameters for creating the actor,

bhv::Function: function implementing the actor's behavior,

args...: arguments to bhv, (without msg)

kwargs...: keyword arguments to bhv.

return a Link to the created actor, a Channel{Message} object.

YAActL.self — Function

self()

Get a Link to yourself from inside an actor.

YAActL.send! — Function

send!(lk::Link, m::Message)

Send a message m to an actor over a link lk.

send!(lks::Tuple{Link,Vararg{Link}}, m::M) where M<:Message
send!(lks::Vector{Link}, m::M) where M<:Message

Send a message m to a Vector or Tuple of Links.

YAActL.become — Function

become(bhv::Function, args...; kwargs...)

Cause yourself to take on a new behavior. Called from inside an actor/behavior.

Arguments

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.become! — Function

become!(lk::Link, bhv::Function, args...; kwargs...)

Cause another actor to assume a new behavior.

Arguments

lk::Link: Link to an actor,

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.stopActor — Function

stop(): an actor terminates.

YAActL.stopActor! — Function

stop!(lk::Link): terminate an actor with link lk.

With become actors can switch their behavior function (e.g. between different state machines) or become! causes them to switch.Thereby an actor can represent a state machine.

Diagnosis

In order to develop actor programs, it is useful to have access to the actor tasks and eventually to their stack traces. You can register! an actor channel to a Vector{Link} in order to get access to the tasks.

YAActL.register! — Function

register!(lks::Vector{Link}, lk::Link)

Register a link lk to a vector of links lks.

Base.istaskfailed — Method

istaskfailed(lk::Link)

Returns true if a task associated with lk has failed.

Base.istaskfailed — Method

istaskfailed(lks::Vector{Link})

Returns true if any task associated with a vector lks of links has failed.

YAActL.taskstate — Function

taskstate(lk::Link)

Return the state (eventually the stacktrace) of a task associated with lk.

1See: The Actor Model on Wikipedia.