Actors

An Actor is a computational entity that, in response to a message it receives, can concurrently:

send a finite number of messages to other actors;

create a finite number of new actors;

designate the behavior to be used for the next message it receives. [1]

YAActL actors are Julia tasks represented by local or remote LINKs, channels over which they receive and send messages [2]. They:

react to those messages,

execute user defined behavior functions when receiving certain messages,

change their behavior upon request,

update their internal state influencing how they behave.

YAActL provides various commands to set, control, trigger and query actor behavior and state.

Start

In the simplest case we start an Actor with a previously implemented behavior and get a link to it.

YAActL.Actor — Function

Actor([lp::LinkParams], bhv::Function, args...; kwargs...)
Actor(pid::Int, bhv::Function, args...; kwargs...)

Create a new actor. Start a task executing repeatedly the behavior bhv. The actor listens to messages msg sent over the returned link and executes bhv(args..., msg, kwargs) for each message. The actor stops if sent Stop().

Arguments

[lp::LinkParams]: optional parameters for creating the actor,

pid::Int: create the actor on process pid,

bhv::Function: function implementing the actor's behavior,

args...: arguments to bhv, (without msg)

kwargs...: keyword arguments to bhv.

Returns

a Link to a locally created actor or

an RLink to a remote actor.

The following code starts an actor with a stack_node behavior function and some arguments to it.

mystack = Actor(stack_node, StackNode(nothing, Link()))

We can now send! messages to mystack or do other operations on it.

Operation

An actor recognizes and operates on predefined messages. Basically there are only two functions to interact with an actor:

send!: send a message to an actor,

receive!: receive a message,

request!: send a message to an actor, block, receive and return the result.

YAActL.send! — Function

send!(lk::LINK, m::Message)

Send a message m to an actor over a LINK lk.

send!(lks::Tuple{LINK,Vararg{LINK}}, m::M) where M<:Message
send!(lks::Vector{LINK}, m::M) where M<:Message

Send a message m to a Vector or Tuple of LINKs.

YAActL.receive! — Function

receive!(lk; timeout=5.0)
receive!(lk, from; timeout=5.0)
receive!(lk, Msg; timeout=5.0)
receive!(lk, Msg, from; timeout=5.0)

Receive a message over a link lk.

If Msg or from are provided, receive! returns only a matching message.

Parameters

lk::LINK: local or remote link over which the message is sent,

Msg::Type{<:Message}: Message type,

from::LINK: local or remote link of sender,

timeout::Real: maximum waiting time in seconds.

Returns

received message or Timeout().

YAActL.request! — Function

request!(lk::LINK, Msg::Type{<:Message}, args...; full=false)

Send a message of type Msg with optional args... to an actor over a given LINK lk, block, receive and return the result.

If full==true return the full Response message.

Message API

Users can interact with actors with explicit messages:

YAActL.Message — Type

Abstract type for messages to actors.

YAActL.Response — Type

Response(y, from::LINK=self())

A Message representing a response to requests.

YAActL.Request — Type

Request(x, from::LINK)

A generic Message for user requests.

YAActL.Timeout — Type

Timeout()

A return value to signal that a timeout has occurred.

If an actor receives a Request message or any other not predefined message, it executes its behavior function with this message as the last argument.

This mechanism can be employed to extend the actor's behavior. A user can implement further message types. For example:

struct Pop <: Message
    customer::Link
end

struct Push{T} <: Message
    content::T
end

User API

The following interface to actors is for common use:

become!: cause an actor to switch its behavior,

call!: call an actor to execute its behavior function and to return the result,

cast!: cause an actor to execute its behavior function,

get: get an actor's internal state,

init!: tell an actor to execute a function at startup,

query: prompt for the result of the last call to the behavior function,

self: get your actor's link,

set!: set the actor's dispatch mode,

stopActor!: terminate an actor,

terminate!: tell an actor to execute a function when it terminates.

update!: update an actor's internal state,

Those functions are wrappers to the predefined messages and to the send! or request! functions. They all involve a communication.

YAActL.become! — Function

become!(lk::LINK, bhv::Function, args...; kwargs...)

Cause an actor to switch its behavior.

Arguments

lk::Link: Link to an actor,

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.call! — Function

call!(lk::LINK, [from::LINK], args...)

Call the lk actors behavior function withargs...and get a [Response](@ref) to thefromchannel. Iffromis omitted,call` blocks and returns the response.

Examples

YAActL.cast! — Function

cast!(lk::LINK, args...)

Cast args... to the lk actor's behavior function. This is an asynchronous message without a response.

Base.get — Function

get(lk::LINK, [from::LINK])

Get a Response message from the lk actor with its internal state to the from channel. If from is omitted, get blocks and returns the response.

YAActL.init! — Function

init!(lk::LINK, f::Function, args...; kwargs...)

Tell an actor lk to execute the init function f with the given arguments at startup.

YAActL.query — Function

query(lk::LINK, [from::LINK])

Query the result of the last call to the behavior function from the lk actor. The Response is sent to the from channel. If from is omitted query blocks and returns the response.

YAActL.update! — Function

update!(lk::LK, args...)

Update the lk actor's internal state with args.... This is an asynchronous message without a response.

If called with Args, it updates the arguments to the behavior function. If Args has keyword arguments, they are merged with existing keyword arguments to the behavior.

YAActL.set! — Function

set!(lk::LINK, dsp::Dispatch)

Set the lk actor's Dispatch to dsp.

YAActL.stopActor! — Function

stopActor!(lk::LINK)

Terminate an actor with link lk. This is an asynchronous message without a response.

YAActL.terminate! — Function

terminate!(lk::LINK, f::Function, args...; kwargs...)

Tell an actor lk to execute a function f when it terminates.

Actors can also operate on themselves or rather they send messages to themselves:

become: and actor switches its own behavior,

self: an actor gets a link to itself,

stopActor: an actor stops.

YAActL.become — Function

become(bhv::Function, args...; kwargs...)

Cause your actor to take on a new behavior. This can only be called from inside an actor/behavior.

Arguments

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.self — Function

self()

Get the LINK of your actor.

YAActL.stopActor — Function

stopActor(): an actor terminates.

Actor State

The actor state is internal and is shared with its environment only for diagnostic purposes. The API functions above are a safe way to access actor state.

1See: The Actor Model on Wikipedia.

2They build on Julia's concurrency primitives @spawn, put! and take! (to/from Channels).