Actor API

The following functions provide a user interface to actors:

Start

YAActL.Actor — Function

Actor([lp::LinkParams], bhv::Function, args...; kwargs...)
Actor(pid::Int, bhv::Function, args...; kwargs...)

Create a new actor. Start a task executing repeatedly the behavior bhv. The actor listens to messages msg sent over the returned link and executes bhv(args..., msg, kwargs) for each message. The actor stops if sent Stop().

Arguments

[lp::LinkParams]: optional parameters for creating the actor,

pid::Int: create the actor on process pid,

bhv::Function: function implementing the actor's behavior,

args...: arguments to bhv, (without msg)

kwargs...: keyword arguments to bhv.

Returns

a Link to a locally created actor or

an RLink to a remote actor.

Messages

Messages in YAActL have a common abstract type. Only a few predefined messages are exported:

YAActL.Message — Type

Abstract type for messages to actors.

YAActL.Response — Type

Response(y, from::LINK=self())

A Message representing a response to requests.

YAActL.Request — Type

Request(x, from::LINK)

A generic Message for user requests.

YAActL.Timeout — Type

Timeout()

A return value to signal that a timeout has occurred.

Actors operate with internal messages. Further messages can be implemented by a user. If an actor receives a message other than an internal one, it passes the message as last argument to the behavior function.

Messages are sent and received using the following primitives:

YAActL.send! — Function

send!(lk::LINK, m::Message)

Send a message m to an actor over a LINK lk.

send!(lks::Tuple{LINK,Vararg{LINK}}, m::M) where M<:Message
send!(lks::Vector{LINK}, m::M) where M<:Message

Send a message m to a Vector or Tuple of LINKs.

YAActL.receive! — Function

receive!(lk; timeout=5.0)
receive!(lk, from; timeout=5.0)
receive!(lk, Msg; timeout=5.0)
receive!(lk, Msg, from; timeout=5.0)

Receive a message over a link lk.

If Msg or from are provided, receive! returns only a matching message.

Parameters

lk::LINK: local or remote link over which the message is sent,

Msg::Type{<:Message}: Message type,

from::LINK: local or remote link of sender,

timeout::Real: maximum waiting time in seconds.

Returns

received message or Timeout().

Actor Control

The following functions control actor behavior and state.

YAActL.become! — Function

become!(lk::LINK, bhv::Function, args...; kwargs...)

Cause an actor to switch its behavior.

Arguments

lk::Link: Link to an actor,

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.cast! — Function

cast!(lk::LINK, args...)

Execute the lk actor's behavior function with args... without sending a response. The returned value of the behavior function can be obtained with query!

YAActL.exit! — Function

exit!(lk::LINK, code=0)

Tell an actor lk to exit. If it has a term function it calls it with code as last argument.

This is an asynchronous message without a response.

YAActL.init! — Function

init!(lk::LINK, f::Function, args...; kwargs...)

Tell an actor lk to execute the init function f with the given arguments at startup and to save the returned value as state variable.

YAActL.update! — Function

update!(lk::LK, args...)

Update the lk actor's internal state with args.... This is an asynchronous message without a response.

It can be called with Args to update the stored arguments to the behavior function. If Args has keyword arguments, they are merged with existing keyword arguments to the behavior function.

YAActL.set! — Function

set!(lk::LINK, dsp::Dispatch)

Set the lk actor's Dispatch to dsp.

YAActL.term! — Function

term!(lk::LINK, f::Function, args...; kwargs...)

Tell an actor lk to execute a function f when it terminates. f must accept a code=0 as last argument. This is added by the actor to args... when it exit!s.

If a behavior function wants to control its own actor, it can use the following functions:

YAActL.become — Function

become(bhv::Function, args...; kwargs...)

Cause your actor to take on a new behavior. This can only be called from inside an actor/behavior.

Arguments

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.self — Function

self()

Get the LINK of your actor.

YAActL.stop — Function

stop(code=0)

Cause your actor to exit with code.

Synchronous Messaging

The following functions send messages to actors causing them to respond. There is one function primitive for this:

YAActL.request! — Function

request!(lk::LINK, Msg::Type{<:Message}, args...; full=false)

Send a message of type Msg with optional args... to an actor over a given LINK lk, block, receive and return the result.

If full==true return the full Response message.

The other functions use it to do specific things:

YAActL.call! — Function

call!(lk::LINK, [from::LINK], args...)

Call the lk actors behavior function withargs...and send its returned value as [Response](@ref) to thefromchannel. Iffromis omitted,call!` blocks and returns the response.

Examples

Base.get! — Function

get!(lk::LINK, [from::LINK])

Get a Response message from the lk actor with its internal state to the from channel. If from is omitted, get! blocks and returns the response.

YAActL.exec! — Function

exec!(lk::LINK, [from::LINK], f::Function, args...; kwargs...)

Tell the lk actor to execute f(args...; kwargs...) and send the returned value as Response to the from channel. If from is omitted, exec! blocks and returns the response value.

YAActL.query! — Function

query!(lk::LINK, [from::LINK])

Query the result of the last call to the behavior function from the lk actor. The Response is sent to the from channel. If from is omitted query! blocks and returns the response.

Note that all those functions block if you don't provide a response channel.