Behaviors

An actor behavior is implemented by different methods of a behavior function. When a message arrives, the actor composes the arguments to its behavior function based on its current state and on the incoming message. This causes a dispatch of one implemented behavior methods.

In the following example we define two behaviors forward! and stack_node. There are two methods for stack_node, dispatching on Push and Pop.

forward!(lk::L, msg::M) where {L<:Link, M<:Message} = send!(lk, msg)

function stack_node(sn::StackNode, msg::Pop)
    isnothing(sn.content) || become(forward!, sn.link)
    send!(msg.customer, Response(sn.content))
end

function stack_node(sn::StackNode, msg::Push)
    P = Actor(stack_node, sn)
    become(stack_node, StackNode(msg.content, P))
end

Behavior Dispatch

Multiple dispatch is arguably one of Julia's strongest features [1]. YAActL actors compose the arguments to the behavior function and thus dispatch one of its implemented methods.

An actor's dispatch mode determines how it

composes the arguments to the behavior function and

uses the function's returned value.

YAActL.Dispatch — Type

Dispatch

Enumeration type for dispatch modes of actors.

Depending on the Dispatch mode the actor composes the arguments to the behavior function:

full: from the Become args... and the msg.x.... This is the default dispatch mode.

state: from the actor state and the msg.x.... In this case the actor updates its state with the result of the behavior function. The result is saved in a Tuple.

The user must implement the needed methods for the behavior dispatch.

Argument Composition

When a message arrives, an actor calls its current behavior function. It composes the function arguments from two parts:

the first part is stored in the actor,

the second part is taken from the message.

Both parts are Tuples and can have multiple arguments and then are composed to form the args... to the behavior function. Then the dispatch happens-

The actor eventually passes keyword arguments kwargs... to the behavior function. But those are not dispatched on.

As user or programmer of an actor system you influence this mechanism in three ways:

First you implement the needed behavior methods with the full number of arguments.

In setting the actor's dispatch mode with set! and by delivering arguments to the behavior function in Actor and become! you determine the state-dependent arguments.

In delivering messages with call!, cast! and send! you determine the message-dependent arguments.

Changing Behavior

An actor's behavior can be set with become! and respectively behavior functions can cause their actor to switch to another behavior.

1see also JuliaCon 2019 | The Unreasonable Effectiveness of Multiple Dispatch | Stefan Karpinski