Behaviors

Actors are created with a behavior function. When a message arrives, they execute their behavior reacting to it.

Therefore the behavior function must take a Message as its last argument.

In the following example we define two behaviors forward! and stack_node. There are two methods for stack_node, dispatching on Push and Pop.

forward!(lk::L, msg::M) where {L<:Link, M<:Message} = send!(lk, msg)

function stack_node(sn::StackNode, msg::Pop)
    isnothing(sn.content) || become(forward!, sn.link)
    send!(msg.customer, Response(sn.content))
end

function stack_node(sn::StackNode, msg::Push)
    P = Actor(stack_node, sn)
    become(stack_node, StackNode(msg.content, P))
end

Dispatching on Messages

Changing Behavior

Actors can change their own behavior or cause other actors to switch behavior.

YAActL.become — Function

become(bhv::Function, args...; kwargs...)

Cause yourself to take on a new behavior. Called from inside an actor/behavior.

Arguments

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.become! — Function

become!(lk::LINK, bhv::Function, args...; kwargs...)

Cause another actor to assume a new behavior.

Arguments

lk::Link: Link to an actor,

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.