Actor State

Actors have an internal state variable where they store behavior and state.

YAActL._ACT — Type

_ACT(lk::LINK)

Internal actor status variable.

Fields

dsp::Dispatch: dispatch mode,

sta::Tuple: the actor's status variable,

res::Tuple: the result of the last behavior execution,

bhv::Func : the behavior function and its internal arguments,

init::Func: the init function and its arguments,

term::Func: the terminate function and its arguments,

link::LINK: the actors (local or remote) link.

see also: Dispatch, Func, LINK

Functions executed by an actor have access to their actor's ACT variable via task_local_storage("ACT").

Normally this is not needed. For accessing the actor's link there is self().

Updating State

To avoid critical race conditions in multithreaded programs actors don't share their internal state with the environment. Only the actor itself is allowed to update its state in a strictly sequential manner by processing message after message.

Other actors or users therefore can cause an actor to update its state only by sending it a message, which is done implicitly by using the API functions.