Links

We send messages to actors and they can send them to others over links. In fact, an actor is only represented by its link which it returns upon creation:

YAActL.Link — Type

A Channel{Message} type for communicating with local actors.

Use buffered channels!

In actor systems we always use buffered links to avoid blocking. Responding on an unbuffered or full link causes an actor to block. Link() creates an unbuffered Channel, use Link(32) or newLink() instead!

Example

julia> response = Link(32)
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.RLink — Type

A RemoteChannel{Link} type for communicating with remote actors.

YAActL.LINK — Constant

A Union{Link, RLink} type for communicating with actors.

Actors correspond with other actors over links. There is a default link for users to communicate with actors.

YAActL.USR — Constant

User remote channel for interacting with actors.

For setting up links explicitly we have the following functions.

YAActL.newLink — Function

newLink(sz::Integer=32)

Create a link of buffer size sz to get responses from actors. Buffer sizes sz < 10 are not allowed.

Example

julia> response = newLink()
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.LinkParams — Type

LinkParams(pid=myid(), size=32; taskref=nothing, spawn=false)

Parameters for setting up an Actor.

pid::Int: process identification,

size::Int: channel buffer size, must be size ≥ 10,

taskref::Union{Nothing, Ref{Task}}: If you need a reference to the created task, pass a Ref{Task} object via the keyword argument taskref.

spawn::Bool: If spawn = true, the Task created may be scheduled on another thread in parallel, equivalent to creating a task via Threads.@spawn.

YAActL.parallel — Function

parallel(size=32; taskref=nothing)

Return LinkParams with spawn=true.

Example

julia> using YAActL, .Threads

julia> myactor = Actor(parallel(), threadid);

julia> call!(myactor)
2