Messages

Messages in YAActL have Message as a common abstract type. Only two predefined messages are exported:

Response: response message type from actor to any synchronous message (requiring a response),

Request: predefined message type for implementing requests to actors.

Messages other than the predefined ones can be implemented by a user.

Sending Functions and Arguments

There are two predefined types for messages with functions and function arguments:

YAActL.Func — Type

Func(f, args...; kwargs...)

A structure for passing a function f and its arguments to an actor.

YAActL.Args — Type

Args(args...; kwargs...)

A structure for updating arguments to an actor's behavior.

Internal Messages

Actors recognize and react to the following predefined messages:

YAActL.Become — Type

Become(x::Func)

An asynchronous Message to an actor to change its behavior.

YAActL.Call — Type

Call(arg, from::LINK)

A synchronous Message to an actor to execute its behavior with arg... and to send the result as a Response message to from.

If the actor is set to state dispatch, it updates its internal state with the result.

YAActL.Cast — Type

Cast(arg)

An asynchronous Message to an actor to execute its behavior with arg... without sending a response.

If the actor is set to state dispatch, it updates its internal state with the result.

YAActL.Diag — Type

Diag(from::LINK)

A synchronous Message to an actor to send a Response message with its internal _ACT variable to from.

YAActL.Get — Type

Get(from::LINK)

A synchronous Message to an actor to send a Response message with its internal state to from.

YAActL.Init — Type

Init(f::Func)

A Message to an actor to execute the given Func and to register it in the _ACT variable.

YAActL.Query — Type

Query(from::LINK)

A synchronous Message to an actor to return the result of the last execution of the behavior function.

YAActL.Set — Type

Set(dsp::Dispatch)
Set(lk::LINK)

An asynchronous Message to an actor to set its Dispatch behavior or its LINK.

YAActL.Stop — Type

Stop()

A Message causing an actor to stop.

YAActL.Term — Type

Term(x::Func)

A Message to an actor to save the given Func and to execute it upon termination.

YAActL.Update — Type

Update(x)

An asynchronous Message to an actor to update its internal state to x. If x is a Args then the arguments to the behavior function are updated.

If an actor receives another subtype of Message, it calls its behavior function with it as last argument.