Usage

Links and messages

We communicate with actors and actors can communicate with each other (and with themselves) over links and messages.

YAActL.Link — Type

Link
Link(size=0)

Is a type Channel{Message} returned by a newly created Actor. Establish a message channel of size used to communicate with actors.

YAActL.LinkParams — Type

LinkParams(size=32; taskref=nothing, spawn=false)

Set the parameters for setting up an Actor. See also: Channel.

Parameters

size::Int: channel buffer size,

taskref::Union{Nothing, Ref{Task}}: If you need a reference to the created task, pass a Ref{Task} object via the keyword argument taskref.

spawn::Bool: If spawn = true, the Task created may be scheduled on another thread in parallel, equivalent to creating a task via Threads.@spawn.

YAActL.parallel — Function

parallel(size=32; taskref=nothing)

Return LinkParams with spawn=true.

YAActL.Message — Type

Abstract type for messages to actors.

YAActL.Stop — Type

message Stop() causes an actor to stop.

To setup an actor system we have to specify the messages. For example:

using YAActL, Printf

struct Pop <: Message
    customer::Link
end

struct Push{T} <: Message
    content::T
end

struct Response{T} <: Message
    content::T
end

Actors and their behaviors

Actors are Julia tasks executing functions as behaviors. If a message arrives, the actor loop passes the message to the behavior function as the last argument. In Julia we want the behavior functions to dispatch on messages. In the following example we define two behaviors forward! and stack_node. There are two methods for stack_node, dispatching on Push and Pop. With become actors can change their behavior. Actors can generate other actors.

forward!(lk::L, msg::M) where {L<:Link, M<:Message} = send!(lk, msg)

function stack_node(sn::StackNode, msg::Pop)
    isnothing(sn.content) || become(forward!, sn.link)
    send!(msg.customer, Response(sn.content))
end

function stack_node(sn::StackNode, msg::Push)
    P = Actor(stack_node, sn)
    become(stack_node, StackNode(msg.content, P))
end

With Actor we can setup actors, with send! we can send them messages, with become! we can cause them to change their behaviors.

mystack = Actor(lk, stack_node, StackNode(nothing, Link()))

Note

If we setup an actor we don't pass the last message argument of the behavior function to Actor.

YAActL.Actor — Function

Actor([lp::LinkParams], bhv::Function, args...; kwargs...)

Create a new actor. Start a task executing repeatedly the behavior bhv. The actor listens to messages msg sent over the returned link and executes bhv(args..., msg, kwargs) for each message. The actor stops if sent Stop().

Arguments

[lp::LinkParams]: optional parameters for creating the actor,

bhv::Function: function implementing the actor's behavior,

args...: arguments to bhv, (without msg)

kwargs...: keyword arguments to bhv.

return a Link to the created actor, a Channel{Message} object.

YAActL.become — Function

become(bhv::Function, args...; kwargs...)

Cause yourself to take on a new behavior. Called from inside an actor/behavior.

Arguments

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.self — Function

self()

Get a Link to yourself from inside an actor.

YAActL.send! — Function

send!(lnk::Link, m::Message)

Send a message m to an actor over a link lnk.

YAActL.become! — Function

become!(lnk::Link, bhv::Function, args...; kwargs...)

Cause another actor to assume a new behavior.

Arguments

lnk::Link: Link to an actor,

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

Internal

YAActL.Become — Type

internal message Become(f::Function, args...; kwargs...) for behavior change.