Usage

Exported

YAActL.Link — Type

Establish a message channel used to communicate with actors.

YAActL.LinkParams — Type

LinkParams(size=32; taskref=nothing, spawn=false)

Set the parameters for establishing a link (a channel) to an actor. Those are forwarded to Channel.

Parameters

size::Int: channel buffer size,

taskref::Union{Nothing, Ref{Task}}: If you need a reference to the created task, pass a Ref{Task} object via the keyword argument taskref.

spawn::Bool: If spawn = true, the Task created may be scheduled on another thread in parallel, equivalent to creating a task via Threads.@spawn.

YAActL.parallel — Function

shortcut for LinkParams with spawn=true.

YAActL.Message — Type

Type for messages to actors.

YAActL.Stop — Type

message Stop() causes an actor to stop.

YAActL.Actor — Function

Actor([lp::LinkParams], bhv::F, args::Vararg{Any, N}; kwargs...) where {F<:Function,N}

Create a new actor. Start a task executing repeatedly the behavior bhv. The actor listens to messages msg sent over the returned link and executes bhv(args..., self, msg, kwargs) for each message. The actor stops if sent Stop().

Arguments

[lp::LinkParams]: optional parameters for creating the link,

bhv::Function: function implementing the actor's behavior,

args...: arguments to bhv, (without msg)

kwargs...: keyword arguments to bhv.

return a link to the created actor, a Channel{Message} object.

YAActL.become — Function

become(bhv::Function, args...; kwargs...)

Cause yourself to take on a new behavior. Called from inside an actor/behavior.

Arguments

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.self — Function

Get a link to yourself from inside an actor.

YAActL.send! — Function

send!(lnk::Link, m::Message)

Send a message m to an actor over a link lnk.

YAActL.become! — Function

become!(lnk::Link, bhv::Function, args...; kwargs...)

Cause another actor to assume a new behavior.

Arguments

lnk::Link: Link to an actor,

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

Internal

YAActL.Become — Type

internal message Become(f::Function, args...; kwargs...) for behavior change.