Usage

Messages

Messages to YAActL actors have a common abstract type. This may seem first a limitation but it enables actors to dispatch their behavior functions on message types. Some basic message types are provided

YAActL.Message — Type

Abstract type for messages to actors.

YAActL.Request — Type

Request(x, u::Link): a message representing a request of x by customer u.

YAActL.Response — Type

Response(y): a message representing a response of y to a request.

YAActL.Stop — Type

Stop(): a message causing an actor to stop.

YAActL.Become — Type

Become(f::Function, args...; kwargs...): internal message for behavior change.

Other message types can be implemented by the user, for example:

struct Pop <: Message
    customer::Link
end

struct Push{T} <: Message
    content::T
end

With dispatch on message types we can easily implement state machines.

Links

We send messages to actors and actors can send them to others over links. In fact, an actor is only represented by its link. If we want a response from an actor, we must send it our own link together with a request.

YAActL.Link — Type

A Channel{Message} type for communicating with actors.

Warn

In actor systems we always use buffered links to avoid blocking. Responding on an unbuffered or full link causes an actor to block. Link() creates an unbuffered Channel, use Link(32) or newLink() instead!

Example

julia> response = Link(32)
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.newLink — Function

newLink(sz::Integer=32)

Create a link of buffer size sz to get responses from actors. Buffer sizes sz < 10 are not allowed.

Example

julia> response = newLink()
Channel{Message}(sz_max:32,sz_curr:0)

YAActL.LinkParams — Type

LinkParams(size=32; taskref=nothing, spawn=false)

Set the parameters for setting up an Actor.

Parameters

size::Int: channel buffer size, must be size ≥ 10,

taskref::Union{Nothing, Ref{Task}}: If you need a reference to the created task, pass a Ref{Task} object via the keyword argument taskref.

spawn::Bool: If spawn = true, the Task created may be scheduled on another thread in parallel, equivalent to creating a task via Threads.@spawn.

persistent::Bool: if persistent = false, the

YAActL.parallel — Function

parallel(size=32; taskref=nothing)

Return LinkParams with spawn=true.

Actors and their behaviors

Actors are Julia tasks executing functions as behaviors. If a message arrives, the actor loop passes the message to the behavior function as the last argument. Therefore a behavior function must take a Message as its last argument.

Often we want the behavior functions to dispatch on messages. In the following example we define two behaviors forward! and stack_node. There are two methods for stack_node, for dispatching on Push and Pop. Actors can change their behavior with become. They can also generate other actors. For example:

# implement behaviors
forward!(lk::L, msg::M) where {L<:Link, M<:Message} = send!(lk, msg)

function stack_node(sn::StackNode, msg::Pop)
    isnothing(sn.content) || become(forward!, sn.link)
    send!(msg.customer, Response(sn.content))
end

function stack_node(sn::StackNode, msg::Push)
    P = Actor(stack_node, sn)
    become(stack_node, StackNode(msg.content, P))
end

With Actor we can setup actors, with send! we can send them messages, with become! we can cause them to change their behaviors.

# setup an actor with a defined behavior
mystack = Actor(lk, stack_node, StackNode(nothing, Link()))

Note

Don't pass the last msg argument of the behavior function to the Actor.

YAActL.Actor — Function

Actor([lp::LinkParams], bhv::Function, args...; kwargs...)

Create a new actor. Start a task executing repeatedly the behavior bhv. The actor listens to messages msg sent over the returned link and executes bhv(args..., msg, kwargs) for each message. The actor stops if sent Stop().

Arguments

[lp::LinkParams]: optional parameters for creating the actor,

bhv::Function: function implementing the actor's behavior,

args...: arguments to bhv, (without msg)

kwargs...: keyword arguments to bhv.

return a Link to the created actor, a Channel{Message} object.

YAActL.self — Function

self()

Get a Link to yourself from inside an actor.

YAActL.send! — Function

send!(lk::Link, m::Message)

Send a message m to an actor over a link lk.

YAActL.become — Function

become(bhv::Function, args...; kwargs...)

Cause yourself to take on a new behavior. Called from inside an actor/behavior.

Arguments

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.become! — Function

become!(lk::Link, bhv::Function, args...; kwargs...)

Cause another actor to assume a new behavior.

Arguments

lk::Link: Link to an actor,

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

YAActL.stop — Function

stop(): an actor terminates.

YAActL.stop! — Function

stop!(lk::Link): terminate an actor with link lk.

By dispatching on messages actors can represent state machines. With become they can switch their behavior between different state machines or become! can cause a switch.